+++ src/main/kotlin/com/lovelysystems/gradle/LovelyGradlePlugin.kt
@@ -31,6 +31,11 @@ open class LovelyPluginExtension(private val project: Project) {
     ) {
         project.dockerProject(repository, stages, platforms, buildPlatforms, dockerFiles)
     }
+
+    fun awsProject(profile: String) {
+        project.awsProject(profile)
+    }
+
 }
 
 class LovelyGradlePlugin : Plugin<Project> {